Refactor CategorySelector into reusable component
